+++ transports/dns/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Update to `libp2p-core` `v0.33.0`.
 
+- Remove implementation of `Clone` on `GenDnsConfig`.
+
 # 0.32.1
 
 - Update to `trust-dns` `v0.21`. See [PR 2543].